requirements.txt
accelerate
numpy
opencv-python
pillow
tensorboard
timm
torch
torchvision
tqdm
